Feat :: ResponseDto Status Long -> Integer

+++ src/main/java/windeath44/server/memorial/domain/presentation/dto/global/ErrorResponseDto.java
@@ -1,7 +1,7 @@
 package windeath44.server.memorial.domain.presentation.dto.global;
 
 public record ErrorResponseDto(
-        Long status,
+        Integer status,
         String message,
         Object data
 ) {
